First pass of config generation: top-level constant attributes
This deals with the simple case of top-level attributes where all of the
instances agree about what the value should be, and thus we can just write
a single constant value into the configuration.

It doesn't yet handle cases where the different instances disagree about
a value, nor any nested blocks that might be present.